requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
langdetect==1.0.9
openai==1.99.9
openai-whisper==20250625
torch==2.8.0
PyMuPDF==1.26.3
pytesseract==0.3.13
Pillow==11.3.0
elevenlabs==2.8.1
